requirements.txt
aiogram
asyncpg
python-dotenv
gspread
